m17: implement SMS packet transmit
Add SMS packet transmit support to OpMode_M17. A new txPacketState()
method handles the TX state machine: encoding the M17 LSF, framing the
SMS payload via PacketFramer, and transmitting the baseband frames.

Introduce SmsTxPacket helper that formats an SMS payload with the 0x05
SMS content type prefix and validates length constraints.

Fix a buffer overflow in PacketFramer::init() by storing the CRC in a
private member variable instead of writing past the caller's buffer.
The CRC bytes are now streamed out by nextFrame() when the frame spans
the data/CRC boundary.

+++ tests/unit/M17_packet.cpp
@@ -560,20 +560,20 @@ TEST_CASE(
 
     // Frame 0: intermediate, counter = 0
     for (size_t i = 0; i < PacketFrame::DATA_SIZE; i++)
-        txFrames[0][i] = static_cast<uint8_t>(i);
+        txFrames[0][i] = i;
     txFrames[0].setEof(false);
     txFrames[0].setCounter(0);
 
     // Frame 1: intermediate, counter = 1
     for (size_t i = 0; i < PacketFrame::DATA_SIZE; i++)
-        txFrames[1][i] = static_cast<uint8_t>(i + PacketFrame::DATA_SIZE);
+        txFrames[1][i] = i + PacketFrame::DATA_SIZE;
     txFrames[1].setEof(false);
     txFrames[1].setCounter(1);
 
     // Frame 2: EOF, counter = number of valid bytes in last chunk (12)
     constexpr uint8_t LAST_BYTES = 12;
     for (size_t i = 0; i < LAST_BYTES; i++)
-        txFrames[2][i] = static_cast<uint8_t>(i + 2 * PacketFrame::DATA_SIZE);
+        txFrames[2][i] = i + 2 * PacketFrame::DATA_SIZE;
     txFrames[2].setEof(true);
     txFrames[2].setCounter(LAST_BYTES);
